+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # REPLACE_GLOBALLY_WITH_NEXT_VERSION
 - PPI-624 - Improved error handling in after order process
+- PPI-628 - Improves payment method choice if PayPal deems the buyer ineligible for certain methods
 - PPI-629 - Fix issue with payment details not showing for APIv1 payments such as PayPal Plus
 
 # 5.0.2